refactor `FPGAEmulator`
