+++ autd3-firmware-emulator/src/fpga/emulator/memory.rs
@@ -1,12 +1,6 @@
 use std::cell::{LazyCell, RefCell};
 
 use autd3_derive::Builder;
-use autd3_driver::{
-    derive::{LoopBehavior, Segment, TransitionMode},
-    ethercat::{DcSysTime, ECAT_DC_SYS_TIME_BASE},
-    firmware::fpga::{Drive, EmitIntensity, GPIOIn, Phase, STMFocus, SilencerTarget},
-};
-use num_integer::Roots;
 
 use crate::FPGAEmulator;
 
@@ -14,22 +8,22 @@ use super::super::params::*;
 
 #[derive(Builder)]
 pub(crate) struct Memory {
-    num_transducers: usize,
+    pub(crate) num_transducers: usize,
     #[get]
-    controller_bram: LazyCell<RefCell<Vec<u16>>>,
+    pub(crate) controller_bram: LazyCell<RefCell<Vec<u16>>>,
     #[get]
-    modulation_bram_0: LazyCell<RefCell<Vec<u16>>>,
+    pub(crate) modulation_bram_0: LazyCell<RefCell<Vec<u16>>>,
     #[get]
-    modulation_bram_1: LazyCell<RefCell<Vec<u16>>>,
+    pub(crate) modulation_bram_1: LazyCell<RefCell<Vec<u16>>>,
     #[get]
-    stm_bram_0: LazyCell<RefCell<Vec<u16>>>,
+    pub(crate) stm_bram_0: LazyCell<RefCell<Vec<u16>>>,
     #[get]
-    stm_bram_1: LazyCell<RefCell<Vec<u16>>>,
+    pub(crate) stm_bram_1: LazyCell<RefCell<Vec<u16>>>,
     #[get]
-    duty_table_bram: LazyCell<RefCell<Vec<u16>>>,
-    tr_pos: LazyCell<Vec<u64>>,
-    sin_table: LazyCell<Vec<u8>>,
-    atan_table: LazyCell<Vec<u8>>,
+    pub(crate) duty_table_bram: LazyCell<RefCell<Vec<u16>>>,
+    pub(crate) tr_pos: LazyCell<Vec<u64>>,
+    pub(crate) sin_table: LazyCell<Vec<u8>>,
+    pub(crate) atan_table: LazyCell<Vec<u8>>,
 }
 
 impl Memory {
@@ -166,362 +160,6 @@ impl FPGAEmulator {
             _ => unreachable!(),
         }
     }
-
-    pub fn fpga_state(&self) -> u16 {
-        self.mem.controller_bram()[ADDR_FPGA_STATE]
-    }
-
-    pub fn assert_thermal_sensor(&mut self) {
-        self.mem.controller_bram_mut()[ADDR_FPGA_STATE] |= 1 << 0;
-    }
-
-    pub fn deassert_thermal_sensor(&mut self) {
-        self.mem.controller_bram_mut()[ADDR_FPGA_STATE] &= !(1 << 0);
-    }
-
-    pub fn is_thermo_asserted(&self) -> bool {
-        (self.mem.controller_bram()[ADDR_FPGA_STATE] & (1 << 0)) != 0
-    }
-
-    pub fn is_force_fan(&self) -> bool {
-        (self.mem.controller_bram()[ADDR_CTL_FLAG] & (1 << CTL_FLAG_FORCE_FAN_BIT)) != 0
-    }
-
-    pub fn gpio_in(&self) -> [bool; 4] {
-        [
-            (self.mem.controller_bram()[ADDR_CTL_FLAG] & (1 << CTL_FLAG_BIT_GPIO_IN_0)) != 0,
-            (self.mem.controller_bram()[ADDR_CTL_FLAG] & (1 << (CTL_FLAG_BIT_GPIO_IN_1))) != 0,
-            (self.mem.controller_bram()[ADDR_CTL_FLAG] & (1 << (CTL_FLAG_BIT_GPIO_IN_2))) != 0,
-            (self.mem.controller_bram()[ADDR_CTL_FLAG] & (1 << (CTL_FLAG_BIT_GPIO_IN_3))) != 0,
-        ]
-    }
-
-    pub fn is_stm_gain_mode(&self, segment: Segment) -> bool {
-        match segment {
-            Segment::S0 => self.mem.controller_bram()[ADDR_STM_MODE0] == STM_MODE_GAIN,
-            Segment::S1 => self.mem.controller_bram()[ADDR_STM_MODE1] == STM_MODE_GAIN,
-            _ => unimplemented!(),
-        }
-    }
-
-    pub fn silencer_update_rate_intensity(&self) -> u16 {
-        self.mem.controller_bram()[ADDR_SILENCER_UPDATE_RATE_INTENSITY]
-    }
-
-    pub fn silencer_update_rate_phase(&self) -> u16 {
-        self.mem.controller_bram()[ADDR_SILENCER_UPDATE_RATE_PHASE]
-    }
-
-    pub fn silencer_completion_steps_intensity(&self) -> u8 {
-        self.mem.controller_bram()[ADDR_SILENCER_COMPLETION_STEPS_INTENSITY] as _
-    }
-
-    pub fn silencer_completion_steps_phase(&self) -> u8 {
-        self.mem.controller_bram()[ADDR_SILENCER_COMPLETION_STEPS_PHASE] as _
-    }
-
-    pub fn silencer_fixed_update_rate_mode(&self) -> bool {
-        (self.mem.controller_bram()[ADDR_SILENCER_FLAG] & SILENCER_FLAG_FIXED_UPDATE_RATE_MODE)
-            == SILENCER_FLAG_FIXED_UPDATE_RATE_MODE
-    }
-
-    pub fn silencer_fixed_completion_steps_mode(&self) -> bool {
-        !self.silencer_fixed_update_rate_mode()
-    }
-
-    pub fn silencer_target(&self) -> SilencerTarget {
-        if (self.mem.controller_bram()[ADDR_SILENCER_FLAG] & SILENCER_FLAG_PULSE_WIDTH)
-            == SILENCER_FLAG_PULSE_WIDTH
-        {
-            SilencerTarget::PulseWidth
-        } else {
-            SilencerTarget::Intensity
-        }
-    }
-
-    pub fn stm_freq_division(&self, segment: Segment) -> u16 {
-        Memory::read_bram_as::<u16>(
-            &self.mem.controller_bram(),
-            match segment {
-                Segment::S0 => ADDR_STM_FREQ_DIV0,
-                Segment::S1 => ADDR_STM_FREQ_DIV1,
-                _ => unimplemented!(),
-            },
-        )
-    }
-
-    pub fn stm_cycle(&self, segment: Segment) -> usize {
-        self.mem.controller_bram()[match segment {
-            Segment::S0 => ADDR_STM_CYCLE0,
-            Segment::S1 => ADDR_STM_CYCLE1,
-            _ => unimplemented!(),
-        }] as usize
-            + 1
-    }
-
-    pub fn sound_speed(&self, segment: Segment) -> u16 {
-        self.mem.controller_bram()[match segment {
-            Segment::S0 => ADDR_STM_SOUND_SPEED0,
-            Segment::S1 => ADDR_STM_SOUND_SPEED1,
-            _ => unimplemented!(),
-        }]
-    }
-
-    pub fn stm_loop_behavior(&self, segment: Segment) -> LoopBehavior {
-        match Memory::read_bram_as::<u16>(
-            &self.mem.controller_bram(),
-            match segment {
-                Segment::S0 => ADDR_STM_REP0,
-                Segment::S1 => ADDR_STM_REP1,
-                _ => unimplemented!(),
-            },
-        ) {
-            0xFFFF => LoopBehavior::infinite(),
-            v => LoopBehavior::finite(v + 1).unwrap(),
-        }
-    }
-
-    pub fn stm_transition_mode(&self) -> TransitionMode {
-        match self.mem.controller_bram()[ADDR_STM_TRANSITION_MODE] as u8 {
-            TRANSITION_MODE_SYNC_IDX => TransitionMode::SyncIdx,
-            TRANSITION_MODE_SYS_TIME => TransitionMode::SysTime(
-                DcSysTime::from_utc(ECAT_DC_SYS_TIME_BASE).unwrap()
-                    + std::time::Duration::from_nanos(Memory::read_bram_as::<u64>(
-                        &self.mem.controller_bram(),
-                        ADDR_STM_TRANSITION_VALUE_0,
-                    )),
-            ),
-            TRANSITION_MODE_GPIO => TransitionMode::GPIO(
-                match Memory::read_bram_as::<u64>(
-                    &self.mem.controller_bram(),
-                    ADDR_STM_TRANSITION_VALUE_0,
-                ) {
-                    0 => GPIOIn::I0,
-                    1 => GPIOIn::I1,
-                    2 => GPIOIn::I2,
-                    3 => GPIOIn::I3,
-                    _ => unreachable!(),
-                },
-            ),
-            TRANSITION_MODE_EXT => TransitionMode::Ext,
-            TRANSITION_MODE_IMMEDIATE => TransitionMode::Immediate,
-            _ => unreachable!(),
-        }
-    }
-
-    pub fn modulation_freq_division(&self, segment: Segment) -> u16 {
-        Memory::read_bram_as::<u16>(
-            &self.mem.controller_bram(),
-            match segment {
-                Segment::S0 => ADDR_MOD_FREQ_DIV0,
-                Segment::S1 => ADDR_MOD_FREQ_DIV1,
-                _ => unimplemented!(),
-            },
-        )
-    }
-
-    pub fn modulation_cycle(&self, segment: Segment) -> usize {
-        self.mem.controller_bram()[match segment {
-            Segment::S0 => ADDR_MOD_CYCLE0,
-            Segment::S1 => ADDR_MOD_CYCLE1,
-            _ => unimplemented!(),
-        }] as usize
-            + 1
-    }
-
-    pub fn modulation_loop_behavior(&self, segment: Segment) -> LoopBehavior {
-        match Memory::read_bram_as::<u16>(
-            &self.mem.controller_bram(),
-            match segment {
-                Segment::S0 => ADDR_MOD_REP0,
-                Segment::S1 => ADDR_MOD_REP1,
-                _ => unimplemented!(),
-            },
-        ) {
-            0xFFFF => LoopBehavior::infinite(),
-            v => LoopBehavior::finite(v + 1).unwrap(),
-        }
-    }
-
-    pub fn modulation_at(&self, segment: Segment, idx: usize) -> u8 {
-        let m = match segment {
-            Segment::S0 => &self.mem.modulation_bram_0()[idx >> 1],
-            Segment::S1 => &self.mem.modulation_bram_1()[idx >> 1],
-            _ => unimplemented!(),
-        };
-        let m = if idx % 2 == 0 { m & 0xFF } else { m >> 8 };
-        m as u8
-    }
-
-    pub fn modulation(&self, segment: Segment) -> Vec<u8> {
-        (0..self.modulation_cycle(segment))
-            .map(|i| self.modulation_at(segment, i))
-            .collect()
-    }
-
-    pub fn mod_transition_mode(&self) -> TransitionMode {
-        match self.mem.controller_bram()[ADDR_MOD_TRANSITION_MODE] as u8 {
-            TRANSITION_MODE_SYNC_IDX => TransitionMode::SyncIdx,
-            TRANSITION_MODE_SYS_TIME => TransitionMode::SysTime(
-                DcSysTime::from_utc(ECAT_DC_SYS_TIME_BASE).unwrap()
-                    + std::time::Duration::from_nanos(Memory::read_bram_as::<u64>(
-                        &self.mem.controller_bram(),
-                        ADDR_MOD_TRANSITION_VALUE_0,
-                    )),
-            ),
-            TRANSITION_MODE_GPIO => TransitionMode::GPIO(
-                match Memory::read_bram_as::<u64>(
-                    &self.mem.controller_bram(),
-                    ADDR_MOD_TRANSITION_VALUE_0,
-                ) {
-                    0 => GPIOIn::I0,
-                    1 => GPIOIn::I1,
-                    2 => GPIOIn::I2,
-                    3 => GPIOIn::I3,
-                    _ => unreachable!(),
-                },
-            ),
-            TRANSITION_MODE_EXT => TransitionMode::Ext,
-            TRANSITION_MODE_IMMEDIATE => TransitionMode::Immediate,
-            _ => unreachable!(),
-        }
-    }
-
-    pub fn req_mod_segment(&self) -> Segment {
-        match self.mem.controller_bram()[ADDR_MOD_REQ_RD_SEGMENT] {
-            0 => Segment::S0,
-            1 => Segment::S1,
-            _ => unreachable!(),
-        }
-    }
-
-    pub fn req_stm_segment(&self) -> Segment {
-        match self.mem.controller_bram()[ADDR_STM_REQ_RD_SEGMENT] {
-            0 => Segment::S0,
-            1 => Segment::S1,
-            _ => unreachable!(),
-        }
-    }
-
-    pub fn pulse_width_encoder_table_at(&self, idx: usize) -> u8 {
-        let v = self.mem.duty_table_bram()[idx >> 1];
-        let v = if idx % 2 == 0 { v & 0xFF } else { v >> 8 };
-        v as u8
-    }
-
-    pub fn pulse_width_encoder_table(&self) -> Vec<u8> {
-        self.mem
-            .duty_table_bram()
-            .iter()
-            .flat_map(|&d| vec![(d & 0xFF) as u8, (d >> 8) as u8])
-            .collect()
-    }
-
-    pub fn debug_types(&self) -> [u8; 4] {
-        [
-            self.mem.controller_bram()[ADDR_DEBUG_TYPE0] as _,
-            self.mem.controller_bram()[ADDR_DEBUG_TYPE1] as _,
-            self.mem.controller_bram()[ADDR_DEBUG_TYPE2] as _,
-            self.mem.controller_bram()[ADDR_DEBUG_TYPE3] as _,
-        ]
-    }
-
-    pub fn debug_values(&self) -> [u16; 4] {
-        [
-            self.mem.controller_bram()[ADDR_DEBUG_VALUE0],
-            self.mem.controller_bram()[ADDR_DEBUG_VALUE1],
-            self.mem.controller_bram()[ADDR_DEBUG_VALUE2],
-            self.mem.controller_bram()[ADDR_DEBUG_VALUE3],
-        ]
-    }
-
-    pub fn num_foci(&self, segment: Segment) -> u8 {
-        self.mem.controller_bram()[match segment {
-            Segment::S0 => ADDR_STM_NUM_FOCI0,
-            Segment::S1 => ADDR_STM_NUM_FOCI1,
-            _ => unimplemented!(),
-        }] as u8
-    }
-
-    pub fn drives(&self, segment: Segment, idx: usize) -> Vec<Drive> {
-        if self.is_stm_gain_mode(segment) {
-            self.gain_stm_drives(segment, idx)
-        } else {
-            self.foci_stm_drives(segment, idx)
-        }
-    }
-
-    fn gain_stm_drives(&self, segment: Segment, idx: usize) -> Vec<Drive> {
-        match segment {
-            Segment::S0 => self.mem.stm_bram_0(),
-            Segment::S1 => self.mem.stm_bram_1(),
-            _ => unimplemented!(),
-        }
-        .iter()
-        .skip(256 * idx)
-        .take(self.mem.num_transducers)
-        .map(|&d| {
-            Drive::new(
-                Phase::new((d & 0xFF) as u8),
-                EmitIntensity::new(((d >> 8) & 0xFF) as u8),
-            )
-        })
-        .collect()
-    }
-
-    fn foci_stm_drives(&self, segment: Segment, idx: usize) -> Vec<Drive> {
-        let bram = match segment {
-            Segment::S0 => self.mem.stm_bram_0(),
-            Segment::S1 => self.mem.stm_bram_1(),
-            _ => unimplemented!(),
-        };
-        let sound_speed = self.sound_speed(segment);
-
-        self.mem
-            .tr_pos
-            .iter()
-            .take(self.mem.num_transducers)
-            .map(|&tr| {
-                let tr_z = ((tr >> 32) & 0xFFFF) as i16 as i32;
-                let tr_x = ((tr >> 16) & 0xFFFF) as i16 as i32;
-                let tr_y = (tr & 0xFFFF) as i16 as i32;
-                let mut intensity = 0x00;
-                let (sin, cos) = (0..self.num_foci(segment) as usize).fold((0, 0), |acc, i| {
-                    let f = unsafe {
-                        (bram[32 * idx + 4 * i..].as_ptr() as *const STMFocus).read_unaligned()
-                    };
-                    let x = f.x();
-                    let y = f.y();
-                    let z = f.z();
-                    let intensity_or_offset = f.intensity();
-                    let offset = if i == 0 {
-                        intensity = intensity_or_offset;
-                        0x00
-                    } else {
-                        intensity_or_offset
-                    };
-
-                    let d2 =
-                        (x - tr_x) * (x - tr_x) + (y - tr_y) * (y - tr_y) + (z - tr_z) * (z - tr_z);
-                    let dist = d2.sqrt() as u32;
-                    let q = ((dist << 14) / sound_speed as u32) as usize;
-                    let q = q + offset as usize;
-                    (
-                        acc.0 + self.mem.sin_table[q % 256] as u16,
-                        acc.1 + self.mem.sin_table[(q + 64) % 256] as u16,
-                    )
-                });
-                let sin = ((sin / self.num_foci(segment) as u16) >> 1) as usize;
-                let cos = ((cos / self.num_foci(segment) as u16) >> 1) as usize;
-                let phase = self.mem.atan_table[(sin << 7) | cos];
-                Drive::new(Phase::new(phase), EmitIntensity::new(intensity))
-            })
-            .collect()
-    }
-
-    pub fn local_tr_pos(&self) -> &[u64] {
-        &self.mem.tr_pos
-    }
 }
 
 #[cfg(test)]
@@ -535,37 +173,5 @@ mod tests {
         let fpga = FPGAEmulator::new(249);
         let addr = (BRAM_SELECT_MOD as u16) << 14;
         fpga.read(addr as _);
-    }
-
-    #[test]
-    fn modulation() {
-        let fpga = FPGAEmulator::new(249);
-        fpga.mem.modulation_bram_0_mut()[0] = 0x1234;
-        fpga.mem.modulation_bram_0_mut()[1] = 0x5678;
-        fpga.mem.controller_bram_mut()[ADDR_MOD_CYCLE0] = 3 - 1;
-        assert_eq!(3, fpga.modulation_cycle(Segment::S0));
-        assert_eq!(0x34, fpga.modulation_at(Segment::S0, 0));
-        assert_eq!(0x12, fpga.modulation_at(Segment::S0, 1));
-        assert_eq!(0x78, fpga.modulation_at(Segment::S0, 2));
-        let m = fpga.modulation(Segment::S0);
-        assert_eq!(3, m.len());
-        assert_eq!(0x34, m[0]);
-        assert_eq!(0x12, m[1]);
-        assert_eq!(0x78, m[2]);
-    }
-
-    #[test]
-    fn is_outputting() {
-        let fpga = FPGAEmulator::new(249);
-        fpga.mem.controller_bram_mut()[ADDR_STM_MODE0] = STM_MODE_GAIN;
-
-        assert!(!fpga.is_outputting());
-
-        fpga.mem.stm_bram_0_mut()[0] = 0xFFFF;
-        assert!(!fpga.is_outputting());
-
-        fpga.mem.modulation_bram_0_mut()[0] = 0xFFFF;
-        fpga.mem.controller_bram_mut()[ADDR_MOD_CYCLE0] = 2 - 1;
-        assert!(fpga.is_outputting());
     }
 }
